backend/requirements.txt
# Transitional pinned dependency artifact kept for compatibility.
# The authoritative dependency source of truth now lives in pyproject.toml.
# Refresh intentionally from pyproject.toml when this file is still needed.
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
click==8.3.1
fastapi==0.109.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
boto3==1.40.74
botocore==1.40.74
azure-identity==1.25.1
azure-storage-blob==12.27.1
google-cloud-storage==2.19.0
Pillow==12.0.0
psycopg[binary]==3.2.10
pydantic==2.12.5
pydantic-core==2.41.5
python-dotenv==1.0.1
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.36.3
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn[standard]==0.27.1
uvloop==0.22.1 ; sys_platform != "win32" and platform_python_implementation == "CPython"
watchfiles==1.1.1
websockets==15.0.1
